Added: double linked list data structure
Added a double linked list data structure using generics
Added some test
Double linked list delete method return the value, which is
bad, it should be return the current deleted node, but
i leave like that to test if it deletes the value.
Find method should return the node, no the value, but i leave like that
to test fast as f.
printForward and printBackward use an generic list to print out the
values of each node.